refactor: 패키지 구조 변경 및 매핑함수 별도로 관리하도록 변경
