+++ app/src/main/java/woowacourse/shopping/data/utils/LocalDateTimeMappter.kt
@@ -1,0 +1,9 @@
+package woowacourse.shopping.data.utils
+
+import java.time.LocalDate
+import java.time.LocalTime
+import java.time.format.DateTimeFormatter
+
+fun String.toLocalTime(): LocalTime = LocalTime.parse(this, DateTimeFormatter.ofPattern("HH:mm:ss"))
+
+fun String.toLocalDate(): LocalDate = LocalDate.parse(this, DateTimeFormatter.ofPattern("yyyy-MM-dd"))
