+++ app/src/main/java/woowacourse/shopping/ui/cart/adapter/CartAdapter.kt
@@ -4,9 +4,9 @@ import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
 import woowacourse.shopping.databinding.ItemCartBinding
-import woowacourse.shopping.ui.CountButtonClickListener
 import woowacourse.shopping.ui.cart.CartItemClickListener
 import woowacourse.shopping.ui.cart.CartUiModel
+import woowacourse.shopping.ui.listener.CountButtonClickListener
 
 class CartAdapter(
     private val cartItemClickListener: CartItemClickListener,
