+++ app/src/main/java/woowacourse/shopping/ui/detail/viewmodel/ProductDetailViewModel.kt
@@ -12,7 +12,7 @@ import woowacourse.shopping.data.product.ProductRepository
 import woowacourse.shopping.data.recentproduct.RecentProductRepository
 import woowacourse.shopping.model.Product
 import woowacourse.shopping.model.ProductWithQuantity
-import woowacourse.shopping.ui.CountButtonClickListener
+import woowacourse.shopping.ui.listener.CountButtonClickListener
 import woowacourse.shopping.ui.utils.MutableSingleLiveData
 import woowacourse.shopping.ui.utils.SingleLiveData
