+++ app/src/main/java/woowacourse/shopping/ui/listener/StartAddProductClickListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.listener
 
 interface StartAddProductClickListener {
     fun addProduct()
