+++ app/src/main/java/woowacourse/shopping/ui/products/adapter/ProductViewHolder.kt
@@ -2,10 +2,10 @@ package woowacourse.shopping.ui.products.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemProductBinding
-import woowacourse.shopping.ui.CountButtonClickListener
+import woowacourse.shopping.ui.listener.AddCartClickListener
+import woowacourse.shopping.ui.listener.CountButtonClickListener
 import woowacourse.shopping.ui.products.ProductItemClickListener
 import woowacourse.shopping.ui.products.ProductWithQuantityUiModel
-import woowacourse.shopping.ui.utils.AddCartClickListener
 
 class ProductViewHolder(
     private val binding: ItemProductBinding,
