+++ app/src/main/java/woowacourse/shopping/ui/utils/bindingAdapters/ImageViewBindingAdapters.kt
@@ -1,7 +1,8 @@
-package woowacourse.shopping.ui.utils
+package woowacourse.shopping.ui.utils.bindingAdapters
 
 import android.widget.ImageView
 import androidx.databinding.BindingAdapter
+import woowacourse.shopping.ui.utils.extension.urlToImage
 
 @BindingAdapter("imageUrl")
 fun ImageView.bindUrlToImage(imageUrl: String?) {
