+++ app/src/main/res/layout/layout_product_count_selection.xml
@@ -14,7 +14,7 @@
 
         <variable
             name="countButtonClickListener"
-            type="woowacourse.shopping.ui.CountButtonClickListener" />
+            type="woowacourse.shopping.ui.listener.CountButtonClickListener" />
 
     </data>
